triphospherpulluper.bat
the main nonsafe file from me its garbage by the way and still considered a trojan so prob run this on a vm or so as virtual machine because when it opens up accessibility on windows do Win+Ctrl+C 2 times and then it will invert the screen. it opens up random websites it moves your mouse cursor it adds a trail to the cursor and inverts the screen so it doesnt destroy the computer tho and doesnt destroy the mbr. Thats everything hope you like it. This is my first repository.